feat: added util function for date parsing
