+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "aries-framework",
-  "version": "0.0.5-unstable.0",
+  "version": "0.0.5-unstable.0-datatransfer.5",
   "license": "Apache-2.0",
   "main": "build/src/index.js",
   "types": "build/src/index.d.ts",
