+++ src/utils/transformers.ts
@@ -1,6 +1,8 @@
 import { Transform, TransformationType } from 'class-transformer'
 import { JsonTransformer } from './JsonTransformer'
 
+import { DateTime } from 'luxon'
+
 /**
  * Decorator that transforms json to and from corresponding record.
  *
@@ -35,4 +37,20 @@ export function RecordTransformer<T>(Class: { new (...args: any[]): T }) {
         return value
     }
   })
+}
+/*
+ * Function that parses date from multiple formats
+ * including SQL formats.
+ */
+
+export function DateParser(value: string): Date {
+  const parsedDate = new Date(value)
+  if (parsedDate instanceof Date && !isNaN(parsedDate.getTime())) {
+    return parsedDate
+  }
+  const luxonDate = DateTime.fromSQL(value)
+  if (luxonDate.isValid) {
+    return new Date(luxonDate.toString())
+  }
+  return new Date()
 }
